Create Bellman_Ford
This program calculates the minimum weight that it takes to get to a "starting" node to an "ending" node. There can be several starting and ending nodes and the program will return the minimum weight from all of the starting and ending nodes. The minimum weight is calculated using a variation of the Bellman-Ford algorithm.

This program takes in three inputs:
paths - This is a list of tuples in the format (node, node, weight) that represent an edge between two nodes
startingNodes - This is a list of nodes that we would like to begin at
endingNodes - This is a list of nodes that we would like to end up at